+++ src/v2/Apps/Search/Routes/Artists/__tests__/SearchResultsArtists.jest.tsx
@@ -1,34 +1,20 @@
 import { ZeroState } from "v2/Apps/Search/Components/ZeroState"
+import { SystemContextProvider } from "v2/Artsy"
 import { PaginationFragmentContainer as Pagination } from "v2/Components/Pagination"
 import { MockBoot } from "v2/DevTools"
 import { mount } from "enzyme"
 import React from "react"
 import { SearchResultsArtistsRoute as SearchResultsArtists } from "../SearchResultsArtists"
 
-jest.mock("v2/Artsy/Router/useRouter", () => ({
-  useRouter: () => ({
-    match: {
-      location: {
-        pathname: "anything",
-      },
-    },
-  }),
-}))
-
-jest.mock("v2/Components/v2/ArtworkFilter/ArtworkFilterContext", () => ({
-  useArtworkFilterContext: () => ({
-    currentlySelectedFilters: () => ({}),
-    filters: {
-      term: "andy",
-    },
-  }),
-}))
+jest.mock("v2/Components/Pagination/useComputeHref")
 
 describe("SearchResultsArtworks", () => {
   const getWrapper = (searchProps: any) => {
     return mount(
       <MockBoot>
-        <SearchResultsArtists {...searchProps} />
+        <SystemContextProvider>
+          <SearchResultsArtists {...searchProps} />
+        </SystemContextProvider>
       </MockBoot>
     )
   }
