+++ src/v2/Components/Pagination/__mocks__/useComputeHref.ts
@@ -1,0 +1,3 @@
+exports.useComputeHref = () => {
+  return () => ""
+}
